revert utils
