+++ .eslintrc.json
@@ -1,9 +1,3 @@
 {
-  "extends": [
-    "plugin:@typescript-eslint/recommended",
-    "next/core-web-vitals",
-    "prettier"
-  ],
-  "parser": "@typescript-eslint/parser",
-  "plugins": ["@typescript-eslint"]
+  "extends": ["next/core-web-vitals", "prettier"]
 }
