+++ package.json
@@ -32,10 +32,6 @@
   "devDependencies": {
     "@tailwindcss/forms": "^0.5.3",
     "@testing-library/react-hooks": "^8.0.1",
-    "@types/node": "^18.14.2",
-    "@types/react": "^18.0.28",
-    "@typescript-eslint/eslint-plugin": "^5.54.0",
-    "@typescript-eslint/parser": "^5.54.0",
     "autoprefixer": "^10.4.13",
     "eslint": "8.26.0",
     "eslint-config-next": "13.0.1",
