+++ utils/webauthn/cbor.js
@@ -1,0 +1,24 @@
+// copy-🍝 from https://github.com/MasterKale/SimpleWebAuthn/blob/33528afe001d4aca62052dce204c0398c3127ffd/packages/server/src/helpers/decodeCbor.ts
+
+export function decodeCborFirst(cbor, input) {
+  try {
+    // throws if there are extra bytes
+    return cbor.decodeFirstSync(input);
+  } catch (err) {
+    const _err = err;
+    // if the error was due to extra bytes, return the unpacked value
+    if (_err.value) {
+      return _err.value;
+    }
+    throw err;
+  }
+}
+
+/**
+ * Intuited from a quick scan of `cbor.decodeFirstSync()` here:
+ *
+ * https://github.com/hildjj/node-cbor/blob/v5.1.0/lib/decoder.js#L189
+ */
+// class CborDecoderError extends Error {
+//   value: any;
+// }
